refactor: move command to root and library into pkg/espflash
