+++ pkg/espflash/chip.go
