+++ pkg/espflash/chip_test.go
