+++ pkg/espflash/errors.go
