+++ pkg/espflash/errors_test.go
