+++ pkg/espflash/flasher_test.go
