+++ pkg/espflash/image_test.go
