+++ pkg/espflash/protocol.go
