+++ pkg/espflash/slip_test.go
