+++ pkg/espflash/target_esp32c6.go
